Rename com.vaadin.server.data to com.vaadin.data.provider
Renamed corresponding test packages too.

Fixes vaadin/framework8-issues#110
