+++ server/src/main/java/com/vaadin/data/provider/DataCommunicator.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.server.data;
+package com.vaadin.data.provider;
 
 import java.io.Serializable;
 import java.util.ArrayList;
