+++ server/src/test/java/com/vaadin/data/HasValueTest.java
@@ -13,7 +13,7 @@
  * License for the specific language governing permissions and limitations under
  * the License.
  */
-package com.vaadin.server.data;
+package com.vaadin.data;
 
 import org.junit.Test;
 import org.mockito.Mockito;
