+++ server/src/test/java/com/vaadin/data/provider/BackendDataProviderTest.java
@@ -1,4 +1,4 @@
-package com.vaadin.server.data.provider;
+package com.vaadin.data.provider;
 
 import java.util.Comparator;
 import java.util.HashMap;
@@ -7,9 +7,9 @@ import java.util.Map;
 import java.util.stream.Collectors;
 import java.util.stream.Stream;
 
+import com.vaadin.data.provider.BackEndDataProvider;
+import com.vaadin.data.provider.SortOrder;
 import com.vaadin.server.SerializablePredicate;
-import com.vaadin.server.data.BackEndDataProvider;
-import com.vaadin.server.data.SortOrder;
 import com.vaadin.shared.data.sort.SortDirection;
 
 public class BackendDataProviderTest extends
