+++ server/src/test/java/com/vaadin/tests/components/grid/GridNoSelectionModelTest.java
@@ -8,7 +8,7 @@ import org.junit.Assert;
 import org.junit.Before;
 import org.junit.Test;
 
-import com.vaadin.server.data.provider.bov.Person;
+import com.vaadin.data.provider.bov.Person;
 import com.vaadin.ui.Grid;
 import com.vaadin.ui.Grid.GridSelectionModel;
 import com.vaadin.ui.Grid.SelectionMode;
