+++ uitest/src/main/java/com/vaadin/tests/components/grid/basics/RefreshDataProvider.java
@@ -17,9 +17,9 @@ package com.vaadin.tests.components.grid.basics;
 
 import java.util.List;
 
+import com.vaadin.data.provider.DataProvider;
+import com.vaadin.data.provider.ListDataProvider;
 import com.vaadin.server.VaadinRequest;
-import com.vaadin.server.data.DataProvider;
-import com.vaadin.server.data.ListDataProvider;
 import com.vaadin.tests.components.AbstractReindeerTestUI;
 import com.vaadin.ui.Button;
 import com.vaadin.ui.Grid;
